동적으로 아이템 설치하도록 변경

+++ Assets/Scripts/Model/Slot.cs
@@ -11,9 +11,9 @@ namespace Model
         public int drink_id;
         public bool has_drink;
         public string incoming_time;
-        public int row;
-        public int column;
-        public int depth;
+        public float row;
+        public float column;
+        public float depth;
 
         public override string ToString()
         {
